requirements.txt
aiogram==2.25.1
python-dotenv==1.0.0
aiosqlite==0.19.0
pydantic==2.5.2 
qrcode[pil]==7.4.2
pyzbar==0.1.9
Pillow==10.0.0